Greg Schlottmann - Sufferfest Challenge Submission
Created a Graph Component per the challenge requirements. Attempted to use Electron, React, and ES6 best practices. This component also responds correctly to window sizing and is property driven. It loads both the file path and height of the component from properties.

Note - I did find and issue with the JSON as it was missing an opening brace and the data is certainly different than the data used to generate the example graph